Allow 'gtk' usage to be configured via feature
Instead of modifying the code to select the textual output or the GTK
output, add a feature allowing this to be selected at build time.

    cargo run

will run the text UI version, and

    cargo run --features use_gtk

will run with the gtk version.

The GTK dependencies are only brought in if the GTK feature is selected.