Improve tests for update_with()
Add to the test cases for update_with() to verify that the mass not only
moves based on the current velocity, but that the velocity also updates
with applied force.

Starting with a mass at ZERO with velocity (1, 1), we check that after a
single time-step with no force, the mass has moved to (1, 1).  Checks
are performed both via PartialEq testing as well as via EPSILON.

Then apply force (2, 3) and verify that the mass has moved to (4, 5) --
since it was at (1, 1) and the velocity went from (1, 1) to (3, 4) due
to the force, the proper final position is (4, 5).

This could be further enahanced to actually test with different masses
(other than 1.0) and also to directly verify the calculated velocity
(rather than inferring it is correct from the position).