Implement Simulator as Trait
This changes Simulator to be a Trait.  The existing Simulator is now
JoeSimulator.  A factory class is required to manufacture new
JoeSimulators in order to be a Trait Object.

Types are excessively specified in this code to make sure that the
typing is always using 'dyn Something' in preparation for
runtime-selectiability of which Simulator model to use.