Implment some useful traits for Point
Implement Add, Sub, and SubAssign for Point.  Interestingly, Point still
needs to have it's own add() method, as the one from the trait can't
be invoked unless the trait is also in scope.  There are other bits of
code that directly call add() which could be fixed, but that's not
important right now.