Replace RefCell with Box for max space coordinates
The original code used RefCell, which allowed for multiple mutable and
immutable references which were checked at runtime (for non-Copy types).
HOWEVER, we don't need multiple references.  We only need an immutable
reference to move into the closure, but then we can re-bind to a mutable
reference to allow us to update the value.