Add colorization of masses; revert Box/RefCell
This uses the magnitude of the velocity of a mass to set the color of
the mass on a gradient that goes from red to green.  Like the max size
of the universe, we track the max speed ever seen and use that for color
scaling.

ALSO: Rever the previous change which moved from RefCell to Box.  The
closure is an Fn closure, not an FnMut, so it can't capture any mutable
variables.  The mutablility of the contents of a box are the same as the
mutablility of the box itself.